requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-sendgrid-v5==0.9.1
django-smtp-ssl==1.0
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
